test_requirements.txt
Django<1.8
django-nose
coverage
django-coverage
django-jasmine
ipdb
flake8
fabric
factory_boy
mock
selenium

django-author>=0.2
south
pillow
django-finegrained-permissions
psycopg2
django-import-export
django-colorful
-e git+https://github.com/PetrDlouhy/django-constance.git@b3f849e0fb#egg=django-constance[database]
django-admin-sortable2
djangorestframework
django-positions
